test: add ui test
