+++ pyproject.toml
@@ -8,6 +8,7 @@ authors = ["Michael Reynaldo Phangtriastu <mikaelphang@gmail.com>"]
 python = "^3.9"
 click = "^8.1.3"
 pydantic = "^1.10.2"
+coverage = "^6.5.0"
 
 [tool.poetry.dev-dependencies]
 black = "^22.8.0"
@@ -17,3 +18,6 @@ pytest = "^7.1.3"
 [build-system]
 requires = ["poetry-core>=1.0.0"]
 build-backend = "poetry.core.masonry.api"
+
+[tool.coverage.run]
+omit = ["*/src/games/*/main.py"]
