+++ tests/tictactoe/bot1.py
@@ -1,0 +1,11 @@
+import click
+
+
+@click.command()
+@click.option("--moves", help="all moves happened")
+def main(moves):
+    print(moves)
+
+
+if __name__ == "__main__":
+    main()
